requirements.txt
numpy
pandas
tensorflow
scikit-learn
Pillow

